+++ src/main/java/racingcar/service/RacingGameService.java
@@ -5,8 +5,8 @@ import racingcar.domain.*;
 import racingcar.dto.request.CarGameRequest;
 import racingcar.dto.response.CarGameResponse;
 import racingcar.dto.response.CarResponse;
-import racingcar.mapper.CarResultDao;
-import racingcar.mapper.PlayResultDao;
+import racingcar.dao.CarResultDao;
+import racingcar.dao.PlayResultDao;
 
 import java.util.Arrays;
 import java.util.List;
